data_handle/requirements.txt
diffusers 
transformers 
torch 
accelerate 
rembg 
pillow
huggingface_hub